presence_app/planning_service/requirements (1).txt
flask==3.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
pyjwt==2.8.0
